install_k8s/gok-agent/v2/controller/requirements.txt
flask
flask-socketio
pika
python-jose
requests
watchdog